move the summary() function to the geojson module
