+++ genet/core.py
@@ -2263,78 +2263,3 @@ class Network:
             slope_dict[link_id] = {'slope': link_slope}
 
         return slope_dict
-
-    def summary(self):
-        """
-        Returns a report with summary statistics for the network and the schedule.
-        """
-        logging.info('Creating a summary report')
-        report = {}
-
-        network_stats = {'Number of network links': nx.number_of_nodes(self.graph),
-                         'Number of network nodes': nx.number_of_edges(self.graph)}
-        report['network_graph_info'] = network_stats
-        report['modes'] = {}
-        report['modes']['Modes on network links'] = self.modes()
-
-        # check for the old format, i.e. long-form attribute notation
-        if len(graph_operations.get_attribute_data_under_key(
-                self.links(), {'attributes': {'osm:way:highway': 'text'}}).values()) == 0:
-            highway_tags = self.link_attribute_data_under_key({'attributes': 'osm:way:highway'})
-            highway_tags = set(itertools.chain.from_iterable(highway_tags.apply(lambda x: persistence.setify(x))))
-        else:
-            highway_tags = self.link_attribute_data_under_key({'attributes': {'osm:way:highway': 'text'}})
-            highway_tags = set(itertools.chain.from_iterable(highway_tags.apply(lambda x: persistence.setify(x))))
-
-        osm_highway_tags = {}
-        for tag in highway_tags:
-            tag_links = self.extract_links_on_edge_attributes(conditions={'attributes': {'osm:way:highway': tag}},
-                                                              mixed_dtypes=True)
-            osm_highway_tags[tag] = len(tag_links)
-        report['osm_highway_tags'] = {'Number of links by tag': osm_highway_tags}
-
-        links_by_mode = {}
-        for mode in self.modes():
-            mode_links = self.extract_links_on_edge_attributes(conditions={'modes': mode}, mixed_dtypes=True)
-            links_by_mode[mode] = len(mode_links)
-        report['modes']['Number of links by mode'] = links_by_mode
-
-        if self.schedule:
-            schedule_stats = {'Number of services': self.schedule.__len__(),
-                              'Number of routes': self.schedule.number_of_routes(),
-                              'Number of stops': len(self.schedule.reference_nodes())}
-            report['schedule_info'] = schedule_stats
-
-            schedule_modes = self.schedule.modes()
-            report['modes']['Modes in schedule'] = schedule_modes
-
-            services_by_modes = {}
-            for mode in schedule_modes:
-                services_by_modes[mode] = len(self.schedule.services_on_modal_condition(mode))
-            report['modes']['Services by mode'] = services_by_modes
-
-            stops_by_modes = {}
-            for mode in schedule_modes:
-                stops_by_modes[mode] = len(self.schedule.stops_on_modal_condition(mode))
-            report['modes']['PT stops by mode'] = stops_by_modes
-
-            report['accessibility_tags'] = {}
-            report['accessibility_tags']['Stops with tag bikeAccessible'] = len(
-                self.schedule.stop_attribute_data({'attributes': 'bikeAccessible'}))
-            bike_accessible_keys = []
-            bike_attribs_dict = self.schedule.stop_attribute_data(
-                {'attributes': 'bikeAccessible'}).to_dict(orient='index')
-            for key in bike_attribs_dict.keys():
-                bike_accessible_keys.append(bike_attribs_dict[key]['attributes']['bikeAccessible'])
-            report['accessibility_tags']['Unique values for bikeAccessible tag'] = set(bike_accessible_keys)
-
-            report['accessibility_tags']['Stops with tag carAccessible'] = len(
-                self.schedule.stop_attribute_data({'attributes': 'carAccessible'}))
-            car_accessible_keys = []
-            car_attribs_dict = self.schedule.stop_attribute_data(
-                {'attributes': 'carAccessible'}).to_dict(orient='index')
-            for key in car_attribs_dict.keys():
-                car_accessible_keys.append(car_attribs_dict[key]['attributes']['carAccessible'])
-            report['accessibility_tags']['Unique values for carAccessible tag'] = set(car_accessible_keys)
-
-        return report
