+++ tests/test_output_geojson.py
@@ -21,6 +21,65 @@ def network(correct_schedule):
     return n
 
 
+@pytest.fixture()
+def network_for_summary_stats():
+    n = Network('epsg:27700')
+    n.add_node('0', attribs={'x': 528704.1425925883, 'y': 182068.78193707118})
+    n.add_node('1', attribs={'x': 528804.1425925883, 'y': 182168.78193707118})
+    n.add_link('link_0', '0', '1', attribs={'length': 123, 'modes': ['car', 'walk'], 'freespeed': 10, 'capacity': 5})
+    n.add_link('link_1', '0', '1', attribs={'length': 123, 'modes': ['bike'],
+                                            'attributes': {'osm:way:highway': 'secondary'}})
+    n.add_link('link_2', '1', '0', attribs={'length': 123, 'modes': ['rail']})
+
+    n.schedule = Schedule(epsg='epsg:27700', services=[
+        Service(id='bus_service',
+                routes=[
+                    Route(id='1', route_short_name='', mode='bus',
+                          stops=[
+                              Stop(id='0', x=529455.7452394223, y=182401.37630677427, epsg='epsg:27700',
+                                   linkRefId='link_1', attributes={'bikeAccessible': 'true',
+                                                                   'accessLinkId_car': '1',
+                                                                   'carAccessible': 'true',
+                                                                   'distance_catchment': '25'}),
+                              Stop(id='1', x=529350.7866124967, y=182388.0201078112, epsg='epsg:27700',
+                                   linkRefId='link_2')],
+                          trips={'trip_id': ['VJ00938baa194cee94700312812d208fe79f3297ee_04:40:00'],
+                                 'trip_departure_time': ['04:40:00'],
+                                 'vehicle_id': ['veh_1_bus']},
+                          arrival_offsets=['00:00:00', '00:02:00'],
+                          departure_offsets=['00:00:00', '00:02:00'],
+                          route=['link_1', 'link_2']),
+                    Route(id='2', route_short_name='route2', mode='bus',
+                          stops=[
+                              Stop(id='0', x=529455.7452394223, y=182401.37630677427, epsg='epsg:27700',
+                                   linkRefId='link_1'),
+                              Stop(id='1', x=529350.7866124967, y=182388.0201078112, epsg='epsg:27700',
+                                   linkRefId='link_2')],
+                          trips={'trip_id': ['1_05:40:00', '2_05:45:00', '3_05:50:00', '4_06:40:00', '5_06:46:00'],
+                                 'trip_departure_time': ['05:40:00', '05:45:00', '05:50:00', '06:40:00', '06:46:00'],
+                                 'vehicle_id': ['veh_2_bus', 'veh_3_bus', 'veh_4_bus', 'veh_5_bus', 'veh_6_bus']},
+                          arrival_offsets=['00:00:00', '00:03:00'],
+                          departure_offsets=['00:00:00', '00:05:00'],
+                          route=['link_1', 'link_2'])
+                ]),
+        Service(id='rail_service',
+                routes=[Route(
+                    route_short_name=r"RTR_I/love\_being//difficult",
+                    mode='rail',
+                    stops=[
+                        Stop(id='RSN', x=-0.1410946, y=51.5231335, epsg='epsg:4326', name=r"I/love\_being//difficult"),
+                        Stop(id='RSE', x=-0.1421595, y=51.5192615, epsg='epsg:4326')],
+                    trips={'trip_id': ['RT1', 'RT2', 'RT3', 'RT4'],
+                           'trip_departure_time': ['03:21:00', '03:31:00', '03:41:00', '03:51:00'],
+                           'vehicle_id': ['veh_7_rail', 'veh_8_rail', 'veh_9_rail', 'veh_10_rail']},
+                    arrival_offsets=['0:00:00', '0:02:00'],
+                    departure_offsets=['0:00:00', '0:02:00']
+                )])
+    ])
+
+    return n
+
+
 def test_saving_values_which_result_in_overflow(tmpdir):
     n = Network('epsg:27700')
     n.add_node('0', attribs={'x': 528704.1425925883, 'y': 182068.78193707118, 's2_id': 7860190995130875979})
@@ -245,3 +304,23 @@ def test_generating_standard_outputs(network, tmpdir):
 
     assert set(os.listdir(os.path.join(tmpdir, 'routing'))) == {'shp_files', 'schedule_network_routes_geodataframe.geojson'}
     assert os.path.exists(tmpdir + '.zip')
+
+
+def test_generating_summary_report(network_for_summary_stats):
+    report = gngeojson.summary(network_for_summary_stats)
+    correct_report = {'network_graph_info':
+                          {'Number of network links': 2,
+                           'Number of network nodes': 3},
+                      'modes': {'Modes on network links': {'car', 'bike', 'rail', 'walk'},
+                                'Number of links by mode': {'car': 1, 'bike': 1, 'rail': 1, 'walk': 1},
+                                'Modes in schedule': {'rail', 'bus'}, 'Services by mode': {'rail': 1, 'bus': 1},
+                                'PT stops by mode': {'rail': 2, 'bus': 2}},
+                      'osm_highway_tags': {'Number of links by tag': {'secondary': 1}},
+                      'schedule_info': {'Number of services': 2,
+                                        'Number of routes': 3,
+                                        'Number of stops': 4},
+                      'accessibility_tags': {'Stops with tag bikeAccessible': 1,
+                                             'Stops with tag carAccessible': 1,
+                                             'Unique values for carAccessible tag': {'true'},
+                                             'Unique values for bikeAccessible tag': {'true'}}}
+    assert_semantically_equal(report, correct_report)
